Rename writer history cache file
