+++ dds/src/implementation/rtps/mod.rs
@@ -2,7 +2,7 @@ pub mod discovery_types;
 pub mod endpoint;
 pub mod entity;
 pub mod group;
-pub mod history_cache;
+pub mod writer_history_cache;
 pub mod participant;
 pub mod reader;
 pub mod reader_locator;
