+++ dds/src/implementation/rtps/reader_locator.rs
@@ -1,5 +1,5 @@
 use super::{
-    history_cache::WriterHistoryCache,
+    writer_history_cache::WriterHistoryCache,
     types::{Locator, SequenceNumber},
 };
